Separate vite to avoid issues with Bun
